# Backend/docs/swagger.yaml
definitions:
...
  controller.UpdateStatusRequest:
    properties:
      status:
        $ref: '#/definitions/domain.MessageStatus'
    required:
    - status
    type: object
...
  domain.ContactMessage:
    properties:
      email:
        maxLength: 255
        type: string
      id:
        type: string
      message:
        type: string
      name:
        maxLength: 150
        type: string
      organization:
        maxLength: 255
        type: string
      phone:
        maxLength: 50
        type: string
      status:
        $ref: '#/definitions/domain.MessageStatus'
      subject:
        maxLength: 300
        type: string
      submitted_at:
        type: string
    required:
    - email
    - message
    - name
    type: object
...
  domain.ErrorDetails:
    properties:
      code:
        type: string
      details:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/domain.ErrorDetails'
      message:
        default: Error
        type: string
      success:
        default: false
        type: boolean
    type: object
...
  domain.MessageStatus:
    enum:
    - unread
    - read
    - replied
    type: string
    x-enum-varnames:
    - StatusUnread
    - StatusRead
    - StatusReplied
...
  domain.SuccessResponse:
    properties:
      data: {}
      message:
        default: Success
        type: string
      success:
        default: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/contacts:
    get:
      description: Retrieve all submitted contact messages for admin review
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ContactMessage'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all contact messages
      tags:
      - admin contact
  /api/admin/contacts/{id}:
    delete:
      description: Delete a contact message by ID
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete contact message
      tags:
      - admin contact
  /api/admin/contacts/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the read/replied status of a contact message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated status payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ContactMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update message status
      tags:
      - admin contact
...
  /api/contacts:
    post:
      consumes:
      - application/json
      description: Submit a contact/inquiry message from the contact form
      parameters:
      - description: Contact message details
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/domain.ContactMessage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ContactMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Submit contact message
      tags:
      - contact
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"